bot/requirements.txt
flask==2.3.2
requests==2.28.1
langchain
google-generativeai
python-dotenv
langchain-google-genai